requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.11.0
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.2
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.118.0
google-auth==2.41.1
google-genai==1.40.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.11.0
Mako==1.3.10
MarkupSafe==3.0.3
motor==3.7.1
openai==2.1.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pymongo==4.15.1
pypdfium2==4.30.0
python-dotenv==1.1.1
python-multipart==0.0.20
regex==2025.9.18
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
websockets==15.0.1
